PPO learning finalized
PPO learning pipeline already finished.
We made the decision to create two optimizers, one
for critic the other for actor. Also, in the minibatch
training, the actor is using the gradients from the detector,
and the critic is detached from the detector.

Replay memory is also improved.

DQN's training was also fixed.